+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project is expected to follow [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+## [Added]
+
+- Add `repeat` .. `until` loop statement.
+
 ## [v0.0.8] - 2026-06-01
 
 ### Added